+++ core/src/edu/byu/rpg/entities/enemies/AI/Attacks/AttackAI.java
@@ -1,5 +1,6 @@
-package edu.byu.rpg.entities.enemies.AI;
+package edu.byu.rpg.entities.enemies.AI.Attacks;
 
+import edu.byu.rpg.entities.enemies.AI.EnemyAI;
 import edu.byu.rpg.entities.enemies.weapons.WeaponType;
 import edu.byu.rpg.entities.enemies.weapons.base.EnemyWeapon;
 import edu.byu.rpg.physics.Body;
