+++ core/src/edu/byu/rpg/entities/enemies/AI/Attacks/FireTrailAI.java
@@ -1,31 +1,27 @@
-package edu.byu.rpg.entities.enemies.AI;
+package edu.byu.rpg.entities.enemies.AI.Attacks;
 
+import edu.byu.rpg.entities.enemies.AI.Attacks.AttackAI;
 import edu.byu.rpg.entities.enemies.weapons.WeaponType;
-import edu.byu.rpg.entities.enemies.weapons.attacks.EnemyBulletWeapon;
-import edu.byu.rpg.entities.enemies.weapons.attacks.EnemyHomingBulletWeapon;
 import edu.byu.rpg.entities.enemies.weapons.base.EnemyWeapon;
 import edu.byu.rpg.physics.Body;
 import edu.byu.rpg.physics.World;
 
 /**
- * Created by Andrew on 2/14/2017.
+ * Created by Andrew on 2/15/2017.
  */
-public class HomingBulletAttackAI implements AttackAI {
-
-    private float attackSpeed = 2.0f;
+public class FireTrailAI implements AttackAI {
+    private float attackSpeed = .5f;
     private float attackDamage = 1.0f;
-    private WeaponType weaponType = WeaponType.HOMING_BULLET;
+    private WeaponType weaponType = WeaponType.TRAIL;
     private EnemyWeapon weapon;
 
-    public HomingBulletAttackAI(){
-    }
+    public FireTrailAI(){}
 
     @Override
     public void scale(float scaleAmount) {
         attackSpeed = attackSpeed * scaleAmount;
         attackDamage = attackDamage * scaleAmount;
         weapon.scale(scaleAmount);
-
     }
 
     @Override
@@ -33,20 +29,19 @@ public class HomingBulletAttackAI implements AttackAI {
 
         if (weapon == null) return;
 
-        //get the X distance to the player
-        float XDir =  world.xDistanceToPlayer(enemyBody);
-        //get the Y distance to the player
-        float YDir = world.yDistanceToPlayer(enemyBody);
+        //if needed to reset cooldown time.
+        weapon.setCooldownTime(0.0f);
+        // get bullet direction and influence by player velocity
+        float XDir = enemyBody.velocity.x;
+        float YDir = enemyBody.velocity.y;
 
         // get center of hitbox
         float x = enemyBody.getCenterX();
         float y = enemyBody.position.y + enemyBody.size.y;
 
-        // set up XDir and YDir for the right and left bullets
-        //fire middle bullet
+        //place firetrail.
         weapon.fire(x, y, XDir, YDir);
     }
-
 
     @Override
     public float getAttackSpeed() {
@@ -66,6 +61,6 @@ public class HomingBulletAttackAI implements AttackAI {
     @Override
     public void setWeapon(EnemyWeapon weapon) {
         this.weapon = weapon;
-    }
 
+    }
 }
