+++ core/src/edu/byu/rpg/entities/enemies/AI/Attacks/HomingBulletAttackAI.java
@@ -1,27 +1,30 @@
-package edu.byu.rpg.entities.enemies.AI;
+package edu.byu.rpg.entities.enemies.AI.Attacks;
 
+import edu.byu.rpg.entities.enemies.AI.Attacks.AttackAI;
 import edu.byu.rpg.entities.enemies.weapons.WeaponType;
-import edu.byu.rpg.entities.enemies.weapons.attacks.EnemyTrailWeapon;
 import edu.byu.rpg.entities.enemies.weapons.base.EnemyWeapon;
 import edu.byu.rpg.physics.Body;
 import edu.byu.rpg.physics.World;
 
 /**
- * Created by Andrew on 2/15/2017.
+ * Created by Andrew on 2/14/2017.
  */
-public class FireTrailAI implements AttackAI {
-    private float attackSpeed = .5f;
+public class HomingBulletAttackAI implements AttackAI {
+
+    private float attackSpeed = 2.0f;
     private float attackDamage = 1.0f;
-    private WeaponType weaponType = WeaponType.TRAIL;
+    private WeaponType weaponType = WeaponType.HOMING_BULLET;
     private EnemyWeapon weapon;
 
-    public FireTrailAI(){}
+    public HomingBulletAttackAI(){
+    }
 
     @Override
     public void scale(float scaleAmount) {
         attackSpeed = attackSpeed * scaleAmount;
         attackDamage = attackDamage * scaleAmount;
         weapon.scale(scaleAmount);
+
     }
 
     @Override
@@ -29,19 +32,20 @@ public class FireTrailAI implements AttackAI {
 
         if (weapon == null) return;
 
-        //if needed to reset cooldown time.
-        weapon.setCooldownTime(0.0f);
-        // get bullet direction and influence by player velocity
-        float XDir = enemyBody.velocity.x;
-        float YDir = enemyBody.velocity.y;
+        //get the X distance to the player
+        float XDir =  world.xDistanceToPlayer(enemyBody);
+        //get the Y distance to the player
+        float YDir = world.yDistanceToPlayer(enemyBody);
 
         // get center of hitbox
         float x = enemyBody.getCenterX();
         float y = enemyBody.position.y + enemyBody.size.y;
 
-        //place firetrail.
+        // set up XDir and YDir for the right and left bullets
+        //fire middle bullet
         weapon.fire(x, y, XDir, YDir);
     }
+
 
     @Override
     public float getAttackSpeed() {
@@ -61,6 +65,6 @@ public class FireTrailAI implements AttackAI {
     @Override
     public void setWeapon(EnemyWeapon weapon) {
         this.weapon = weapon;
-
     }
+
 }
