+++ core/src/edu/byu/rpg/entities/enemies/AI/Attacks/ThreeBulletAttackAI.java
@@ -1,7 +1,7 @@
-package edu.byu.rpg.entities.enemies.AI;
+package edu.byu.rpg.entities.enemies.AI.Attacks;
 
+import edu.byu.rpg.entities.enemies.AI.Attacks.AttackAI;
 import edu.byu.rpg.entities.enemies.weapons.WeaponType;
-import edu.byu.rpg.entities.enemies.weapons.attacks.EnemyBulletWeapon;
 import edu.byu.rpg.entities.enemies.weapons.base.EnemyWeapon;
 import edu.byu.rpg.physics.Body;
 import edu.byu.rpg.physics.World;
