+++ core/src/edu/byu/rpg/entities/enemies/AI/Movement/FollowMovementAI.java
@@ -1,0 +1,24 @@
+package edu.byu.rpg.entities.enemies.AI.Movement;
+
+import edu.byu.rpg.physics.Body;
+import edu.byu.rpg.physics.World;
+
+/**
+ * Created by Andrew on 2/20/2017.
+ */
+public class FollowMovementAI implements MovementAI {
+    @Override
+    public void scale(float scaleAmount) {
+        
+    }
+
+    @Override
+    public void move(Body enemyBody, World world) {
+
+    }
+
+    @Override
+    public float getMovementSpeed() {
+        return 0;
+    }
+}
