+++ core/src/edu/byu/rpg/entities/enemies/Scarab.java
@@ -4,8 +4,8 @@ import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import edu.byu.rpg.RpgGame;
 import edu.byu.rpg.entities.base.Actor;
 import edu.byu.rpg.entities.effects.Shadow;
-import edu.byu.rpg.entities.enemies.AI.AttackAI;
-import edu.byu.rpg.entities.enemies.AI.MovementAI;
+import edu.byu.rpg.entities.enemies.AI.Attacks.AttackAI;
+import edu.byu.rpg.entities.enemies.AI.Movement.MovementAI;
 import edu.byu.rpg.entities.enemies.weapons.WeaponType;
 import edu.byu.rpg.entities.enemies.weapons.attacks.EnemyBulletWeapon;
 import edu.byu.rpg.entities.enemies.weapons.attacks.EnemyHomingBulletWeapon;
