+++ core/src/edu/byu/rpg/screens/PlayScreen.java
@@ -8,17 +8,12 @@ import com.badlogic.gdx.maps.tiled.objects.TiledMapTileMapObject;
 import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
 import com.badlogic.gdx.math.Rectangle;
 import edu.byu.rpg.RpgGame;
-import edu.byu.rpg.audio.AudioManager;
 import edu.byu.rpg.entities.base.Solid;
-import edu.byu.rpg.entities.enemies.AI.*;
-import edu.byu.rpg.entities.enemies.AI.BulletAttackAI;
-import edu.byu.rpg.entities.enemies.AI.FireTrailAI;
-import edu.byu.rpg.entities.enemies.AI.RandomMovementAI;
-import edu.byu.rpg.entities.enemies.AI.ThreeBulletAttackAI;
+import edu.byu.rpg.entities.enemies.AI.Attacks.AttackType;
+import edu.byu.rpg.entities.enemies.AI.Movement.MovementType;
 import edu.byu.rpg.entities.enemies.MonsterType;
 import edu.byu.rpg.entities.enemies.controllers.AIController;
 import edu.byu.rpg.entities.enemies.controllers.EnemyController;
-import edu.byu.rpg.entities.enemies.weapons.attacks.EnemyBulletWeapon;
 import edu.byu.rpg.entities.player.Player;
 import edu.byu.rpg.physics.Body;
 import edu.byu.rpg.physics.World;
